+++ src/oracleFactory/OracleReader.ts
@@ -20,7 +20,7 @@ import {
   DataType,
   ParamSetCID,
   OracleID,
-} from '../types/public-types.js';
+} from '../types/index.js';
 import { readOracle } from './readOracle.js';
 
 /**
@@ -51,6 +51,7 @@ class IExecOracleReader {
     ethProviderOrNetwork:
       | Web3ReadOnlyProvider
       | Web3SignerProvider
+      | AbstractProvider
       | Eip1193Provider
       | string
       | number = 134,