+++ src/utils/errors.ts
@@ -1,17 +1,43 @@
+import { ApiCallError } from 'iexec/errors';
 import { ValidationError } from 'yup';
+import { SafeObserver } from './reactive.js';
 
-class WorkflowError extends Error {
-  originalError?: Error;
+export const updateErrorMessage = 'Failed to update oracle';
 
-  constructor(message: string, originalError?: Error) {
-    super(message);
+export class WorkflowError extends Error {
+  isProtocolError: boolean;
+
+  constructor({
+    message,
+    errorCause,
+    isProtocolError = false,
+  }: {
+    message: string;
+    errorCause: Error;
+    isProtocolError?: boolean;
+  }) {
+    super(message, { cause: errorCause });
     this.name = this.constructor.name;
-    if (originalError) {
-      this.originalError = originalError;
-    }
+    this.isProtocolError = isProtocolError;
   }
 }
 
-class NoValueError extends Error {}
+export function handleIfProtocolError(
+  error: Error,
+  observer: SafeObserver<unknown>
+) {
+  if (error instanceof ApiCallError) {
+    observer.error(
+      new WorkflowError({
+        message:
+          "A service in the iExec protocol appears to be unavailable. You can retry later or contact iExec's technical support for help.",
+        errorCause: error,
+        isProtocolError: true,
+      })
+    );
+  }
+}
 
-export { ValidationError, WorkflowError, NoValueError };
+export class NoValueError extends Error {}
+
+export { ValidationError };