+++ tests/e2e/IExecOracleFactory/readOracle.e2e.test.ts
@@ -1,16 +1,9 @@
-import { Wallet } from 'ethers';
-import { utils } from 'iexec';
-import { IExecOracleFactory } from '../../../src/index.js';
+import { IExecOracleReader } from '../../../src/index.js';
 
 describe('readOracle', () => {
   test('standard - from paramSet dataType: "boolean"', async () => {
-    const ethProvider = utils.getSignerFromPrivateKey(
-      'bellecour',
-      Wallet.createRandom().privateKey
-    );
-    const factoryWithoutOption = new IExecOracleFactory(ethProvider);
-
-    const res = await factoryWithoutOption.readOracle({
+    const oracleReader = new IExecOracleReader(134);
+    const res = await oracleReader.readOracle({
       JSONPath: '$.ok',
       body: '',
       dataType: 'boolean',
@@ -25,13 +18,8 @@ describe('readOracle', () => {
   });
 
   test('standard - from paramSet dataType: "number"', async () => {
-    const ethProvider = utils.getSignerFromPrivateKey(
-      'bellecour',
-      Wallet.createRandom().privateKey
-    );
-    const factoryWithoutOption = new IExecOracleFactory(ethProvider);
-
-    const res = await factoryWithoutOption.readOracle({
+    const oracleReader = new IExecOracleReader(134);
+    const res = await oracleReader.readOracle({
       JSONPath: "$['ethereum']['usd']",
       body: '',
       dataType: 'number',
@@ -46,13 +34,8 @@ describe('readOracle', () => {
   });
 
   test('standard - from paramSet dataType: "string"', async () => {
-    const ethProvider = utils.getSignerFromPrivateKey(
-      'bellecour',
-      Wallet.createRandom().privateKey
-    );
-    const factoryWithoutOption = new IExecOracleFactory(ethProvider);
-
-    const res = await factoryWithoutOption.readOracle({
+    const oracleReader = new IExecOracleReader(134);
+    const res = await oracleReader.readOracle({
       JSONPath: '$.version',
       body: '',
       dataType: 'string',
@@ -67,13 +50,8 @@ describe('readOracle', () => {
   });
 
   test('standard - from CID', async () => {
-    const ethProvider = utils.getSignerFromPrivateKey(
-      'bellecour',
-      Wallet.createRandom().privateKey
-    );
-    const factoryWithoutOption = new IExecOracleFactory(ethProvider);
-
-    const res = await factoryWithoutOption.readOracle(
+    const oracleReader = new IExecOracleReader(134);
+    const res = await oracleReader.readOracle(
       'Qmb1JLTVp4zfRMPaori9htzzM9D3B1tG8pGbZYTRC1favA'
     );
     const { value, date } = res;
@@ -82,13 +60,8 @@ describe('readOracle', () => {
   });
 
   test('standard - from oracleId (default dataType)', async () => {
-    const ethProvider = utils.getSignerFromPrivateKey(
-      'bellecour',
-      Wallet.createRandom().privateKey
-    );
-    const factoryWithoutOption = new IExecOracleFactory(ethProvider);
-
-    const res = await factoryWithoutOption.readOracle(
+    const oracleReader = new IExecOracleReader(134);
+    const res = await oracleReader.readOracle(
       '0xf0f370ad33d1e3e8e2d8df7197c40f62b5bc403553b103858359687491234491'
     );
     const { value, date } = res;
@@ -97,13 +70,8 @@ describe('readOracle', () => {
   });
 
   test('standard - from oracleId (dataType number)', async () => {
-    const ethProvider = utils.getSignerFromPrivateKey(
-      'bellecour',
-      Wallet.createRandom().privateKey
-    );
-    const factoryWithoutOption = new IExecOracleFactory(ethProvider);
-
-    const res = await factoryWithoutOption.readOracle(
+    const oracleReader = new IExecOracleReader(134);
+    const res = await oracleReader.readOracle(
       '0x31172fe38a7be8a62fa4882d3a5b5cf7da13fa6ad5b144a0c2f35b559bbba14f',
       { dataType: 'number' }
     );
@@ -113,13 +81,8 @@ describe('readOracle', () => {
   });
 
   test('standard - from oracleId (dataType string)', async () => {
-    const ethProvider = utils.getSignerFromPrivateKey(
-      'bellecour',
-      Wallet.createRandom().privateKey
-    );
-    const factoryWithoutOption = new IExecOracleFactory(ethProvider);
-
-    const res = await factoryWithoutOption.readOracle(
+    const oracleReader = new IExecOracleReader(134);
+    const res = await oracleReader.readOracle(
       '0x9fc5c194d4898197e535060b54256435fda773ae59c93cf88be84bce1ca4ce3e',
       { dataType: 'string' }
     );
@@ -129,13 +92,8 @@ describe('readOracle', () => {
   });
 
   test('standard - from oracleId (dataType boolean)', async () => {
-    const ethProvider = utils.getSignerFromPrivateKey(
-      'bellecour',
-      Wallet.createRandom().privateKey
-    );
-    const factoryWithoutOption = new IExecOracleFactory(ethProvider);
-
-    const res = await factoryWithoutOption.readOracle(
+    const oracleReader = new IExecOracleReader(134);
+    const res = await oracleReader.readOracle(
       '0xccf7d910abf22fbeeef17f861b5cf9abb9543e48ee502285f7df53c63296ce21',
       { dataType: 'boolean' }
     );