+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.2.0] 2024-09-17
+
+### Added
+
+- Support for ethers `AbstractProvider` and `AbstractSigner` in constructors
+- Export more types
+
+### Changed
+
+- Upgraded to iexec ^8.10.0
+- [BREAKING] Ship ES2022 JavaScript instead of es2015 (aka es6) in order to support `errorCause` optional field in `Error`:
+  - Minimum browser versions: <https://gist.github.com/Julien-Marcou/156b19aea4704e1d2f48adafc6e2acbf>
+  - Minimum Node.js version: 18
+- Upgrade typescript version
+- Changed `createOracle`, `updateOracle` and `readOracle` error handling:
+  - Distinguish iExec protocol errors from other errors
+  - Store original error as the error errorCause
+- [BREAKING] Removed `originalError` from `WorkflowError`
+- Fixed supported target chains for cross-chain (134 is not included)
+
+### Removed
+
+- Removed support for mumbai cross-chain oracles
+
 ## [2.1.0] 2024-03-15
 
 ### Removed